refactor: Move cwrap into internal package
