+++ internal/msg_types.c
