+++ internal/msg_types.h
