+++ node/node.go
@@ -6,7 +6,7 @@ package node
 // #include "rcl/node.h"
 import "C"
 import (
-	"github.com/richardrigby/rclgo/cwrap"
+	cwrap "github.com/richardrigby/rclgo/internal"
 	"github.com/richardrigby/rclgo/rcl"
 	"github.com/richardrigby/rclgo/types"
 )
