+++ types/msg_std_msgs_uint32.go
@@ -5,13 +5,13 @@ package types
 /////////////////////////////////////////////////////
 // #cgo CFLAGS: -I/opt/ros/eloquent/include
 // #cgo LDFLAGS: -L/opt/ros/eloquent/lib -Wl,-rpath=/opt/ros/eloquent/lib -lrcl -lrosidl_generator_c -lrosidl_typesupport_c -lstd_msgs__rosidl_generator_c -lstd_msgs__rosidl_typesupport_c
-// #include "../cwrap/msg_types.h"
+// #include "../internal/msg_types.h"
 import "C"
 
 import (
 	"unsafe"
 
-	"github.com/richardrigby/rclgo/cwrap"
+	cwrap "github.com/richardrigby/rclgo/internal"
 )
 
 type StdMsgsUInt32 struct {
